requirements.txt
torch
torchvision
opencv-python-headless
mediapipe
numpy
pillow
dlib